Move translatable nodes to a public static method to avoid problems in tests
Explanation: The translatable node array is just a configuration and
there's nothing to really test except that the method is actually
returning an array.

Here, one test was failing just because we were mocking the filter `wpml_cornerstone_modules_to_translate` and we
needed to pass the expected value. But the test has really nothing to do
with the change we applied.

The configuration should be tested as part of a CC test.

FTR, I already applied the same change in Elementor
(https://github.com/OnTheGoSystems/wpml-page-builders-elementor/commit/a0d70c2d8bdb56f778311c459ac53aeaf461222e)

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlcore-7232